+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/LeftChatMemberImpl.kt
@@ -1,13 +1,14 @@
 package dev.inmo.tgbotapi.types.chat.member
 
 import dev.inmo.tgbotapi.types.*
+import dev.inmo.tgbotapi.types.chat.PreviewUser
 import dev.inmo.tgbotapi.types.chat.User
 import kotlinx.serialization.*
 
 @Serializable
 data class LeftChatMemberImpl(
     @SerialName(userField)
-    override val user: User
+    override val user: PreviewUser
 ) : LeftChatMember {
     @SerialName(statusField)
     @Required