+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/RestrictedChatMember.kt
@@ -2,6 +2,7 @@ package dev.inmo.tgbotapi.types.chat.member
 
 import dev.inmo.tgbotapi.abstracts.types.UntilDate
 import dev.inmo.tgbotapi.types.*
+import dev.inmo.tgbotapi.types.chat.PreviewUser
 import dev.inmo.tgbotapi.types.chat.User
 import kotlinx.serialization.Serializable
 
@@ -17,7 +18,7 @@ sealed interface RestrictedChatMember : ChatMember, UntilDate {
             )
         )
         operator fun invoke(
-            user: User,
+            user: PreviewUser,
             untilDate: TelegramDate? = null,
             isMember: Boolean = false,
             canSendMessages: Boolean = false,