+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/RestrictedMemberChatMember.kt
@@ -2,6 +2,7 @@ package dev.inmo.tgbotapi.types.chat.member
 
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.chat.ChatPermissions
+import dev.inmo.tgbotapi.types.chat.PreviewUser
 import dev.inmo.tgbotapi.types.chat.User
 import kotlinx.serialization.*
 
@@ -12,7 +13,7 @@ import kotlinx.serialization.*
 @Serializable
 data class RestrictedMemberChatMember(
     @SerialName(userField)
-    override val user: User,
+    override val user: PreviewUser,
     @SerialName(untilDateField)
     override val untilDate: TelegramDate? = null,
     @SerialName(isMemberField)